feat: 게시물 관련 예외 처리

+++ src/main/java/com/leets/team2/xclone/domain/post/service/PostService.java
@@ -6,6 +6,7 @@ import com.leets.team2.xclone.domain.member.service.MemberService;
 import com.leets.team2.xclone.domain.post.dto.PostDTO;
 import com.leets.team2.xclone.domain.post.entity.Post;
 import com.leets.team2.xclone.domain.post.repository.PostRepository;
+import com.leets.team2.xclone.exception.PostNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -27,13 +28,16 @@ public class PostService {
 
     public Post updatePost(Long postId,PostDTO postDTO){
         Post post=postRepository.findById(postId)
-                .orElseThrow(()->new NoSuchElementException("해당 게시물이 없습니다."));
+                .orElseThrow(PostNotFoundException::new);
         post.setTitle(postDTO.getTitle());
         post.setContent(postDTO.getContent());
         return postRepository.save(post);
     }
 
     public void deletePost(Long postId){
+        if(!postRepository.existsById(postId)){//게시물이 없을 경우 예외 발생
+            throw new PostNotFoundException();
+        }
         postRepository.deleteById(postId);//해당 게시물 아이디의 게시물 삭제
     }
 }
